requirements.txt
streamlit
tensorflow
opencv-python
pillow
reportlab
numpy